@@ -14,12 +14,12 @@
 /** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
 const sidebars = {
   // By default, Docusaurus generates a sidebar from the docs folder structure
-  tutorialSidebar: [{ type: 'autogenerated', dirName: '.' }],
+  tutorialSidebar: [{type: 'autogenerated', dirName: '.'}],
 
   // But you can create a sidebar manually
   /*
   tutorialSidebar: [
-    'about',
+    'intro',
     'hello',
     {
       type: 'category',
@@ -30,4 +30,4 @@ const sidebars = {
    */
 };
 
-module.exports = sidebars;
+export default sidebars;
